--- a/src/libs/actions/OnyxDerived/configs/reportAttributes.ts
+++ b/src/libs/actions/OnyxDerived/configs/reportAttributes.ts
@@ -92,7 +92,6 @@ export default createOnyxDerivedValueConfig({
             }
 
             const reportActionsList = reportActions?.[`${ONYXKEYS.COLLECTION.REPORT_ACTIONS}${report.reportID}`];
-            // console.log('[p] transactionViolations', transactionViolations, reports);
             const {hasAnyViolations, requiresAttention, isReportArchived} = generateReportAttributes({
                 report,
                 reportActions,